drivers: dma-removed: use relaxed memory attributes
For mapping, dma-removed uses ioremap which on ARM translates
to Device nGnRE. But since the users of dma-removed maps RAM
and does not have restrictions on gathering, reordering, and
speculation while mapped, use ioremap_wc instead. This is
observed to perform better. This patch fixes the variable
address space markings too.

Change-Id: Ic73f38ceb446f19a86cb3f64cbf029a61de6d350
Signed-off-by: Vinayak Menon <vinmenon@codeaurora.org>
Signed-off-by: SJD Ayy <1sjd@disroot.org>