staging: sync: Remove fence names by default
Copying fence debug names for frequent users of sync fences causes a
measurably significant portion of CPU time to be spent on the strlcpy.
Additionally, the sync_fence struct is allocated very frequently, so
wasting space on an unneeded 64-byte name is suboptimal.

Remove the 64-byte name from the struct to reduce its size considerably
and eliminate unnecessary operations on fence names.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Zile995 <stefan.zivkovic995@gmail.com>
Signed-off-by: dracarys18 <karthihegde010@gmail.com>
Signed-off-by: SJD Ayy <3xp1r4t3d@gmail.com>