net: ipc_router: Fix for system hang during graceful shutdown
During the graceful shutdown(using reboot command)
qmi client will get the remove server message
and client will release the handle. Immediately
after release client will lookup for service
and it is succeeding. So client will try to
communicate with service and hit on error callback.
Again same process will continue which blocks the
graceful shutdown.

The issue is due to server blocked on mutex to get the
access to send data over usb to remote processor. The
mutex is acquired by other service and blocked on USB
write API which will cause shutdown failure. So this
is workaround which will help to resolve the issue or
reduce the frequency of issue. This need to be reverted
once real fix identified.

Change-Id: I927987b2cc4b4717ff4f709fcd0b7e228a9cf5cd
Signed-off-by: Arun Prakash <app@codeaurora.org>
Signed-off-by: SJD Ayy <3xp1r4t3d@gmail.com>